requirements.txt
Django==3.1.3
djangorestframework==3.12.1
PyJWT==1.7.1
drf-nested-routers==0.92.1
Pillow==8.0.1
